refactor: extract run report and test helpers for #28
